+++ include/server/server.hpp
@@ -951,7 +951,7 @@ namespace awh {
 			 * @return    результат установки MTU сетевого интерфейса
 			 *
 			 */
-			virtual bool setMaximumTransmissionUnit(const event::id_t eid, const uint16_t mtu) const noexcept;
+			virtual bool setMaximumTransmissionUnit(const event::id_t eid, const uint32_t mtu) const noexcept;
 		public:
 			/**
 			 * @brief Метод получения режима трансляции пакетов сервера или клиента
