+++ include/units/quic.hpp
@@ -823,7 +823,7 @@ namespace awh {
 				 * @return    результат установки MTU сетевого интерфейса
 				 *
 				 */
-				bool setMaximumTransmissionUnit(const event::id_t eid, const uint16_t mtu) const noexcept;
+				bool setMaximumTransmissionUnit(const event::id_t eid, const uint32_t mtu) const noexcept;
 				/**
 				 * @brief Метод получения режима обнаружения максимального размера пакета (MTU)
 				 *
@@ -1859,7 +1859,7 @@ namespace awh {
 				 * @return    результат установки MTU сетевого интерфейса
 				 *
 				 */
-				bool setMaximumTransmissionUnit(const event::id_t eid, const uint16_t mtu) const noexcept;
+				bool setMaximumTransmissionUnit(const event::id_t eid, const uint32_t mtu) const noexcept;
 				/**
 				 * @brief Метод получения режима обнаружения максимального размера пакета (MTU)
 				 *
