+++ src/units/client.cpp
@@ -822,7 +822,7 @@ uint16_t awh::unit::Client::getMaximumTransmissionUnit(const event::id_t eid) co
  * @return    результат установки MTU сетевого интерфейса
  *
  */
-bool awh::unit::Client::setMaximumTransmissionUnit(const event::id_t eid, const uint16_t mtu) const noexcept {
+bool awh::unit::Client::setMaximumTransmissionUnit(const event::id_t eid, const uint32_t mtu) const noexcept {
 	// Если событие клиента является актуальным
 	if(this->isActual(eid))
 		// Выполняем установку MTU сетевого интерфейса для события клиента
