+++ src/units/server.cpp
@@ -1374,7 +1374,7 @@ uint16_t awh::unit::Server::getMaximumTransmissionUnit(const event::id_t eid) co
  * @return    результат установки MTU сетевого интерфейса
  *
  */
-bool awh::unit::Server::setMaximumTransmissionUnit(const event::id_t eid, const uint16_t mtu) const noexcept {
+bool awh::unit::Server::setMaximumTransmissionUnit(const event::id_t eid, const uint32_t mtu) const noexcept {
 	// Если событие сервера является актуальным
 	if(this->isActual(eid))
 		// Выполняем установку MTU сетевого интерфейса для события сервера
